does refactoring of login content
